Added new Spine demo - FlxSpineLoader
This demo illustrates the list of the Spine object animations.
And you can play with different mix of animations in real example.
Also you can control time scale data for different freeze features in
your project.

This demo doesn't have loader feature from FileBrowser, but it's the
next step of the demo.